requirements.txt
python-telegram-bot
google-generativeai
requests
Pillow
telegram
flask